Consolidate Breadcrumbs components
