+++ src/components/Breadcrumbs/BreadcrumbContainer.js
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { uiColors } from '@leafygreen-ui/palette';
 import { css } from '@emotion/react';
-import Link from './Link';
-import { NavigationContext } from './navigation-context';
-import { formatText } from '../utils/format-text';
-import { reportAnalytics } from '../utils/report-analytics';
+import Link from '../Link';
+import { NavigationContext } from '../navigation-context';
+import { formatText } from '../../utils/format-text';
+import { reportAnalytics } from '../../utils/report-analytics';
 
 const activeColor = css`
   color: ${uiColors.gray.dark3};
