+++ src/components/Breadcrumbs/BreadcrumbSchema.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withPrefix } from 'gatsby';
 import { Helmet } from 'react-helmet';
-import { useSiteMetadata } from '../hooks/use-site-metadata';
-import { assertTrailingSlash } from '../utils/assert-trailing-slash';
-import { baseUrl } from '../utils/dotcom';
+import { useSiteMetadata } from '../../hooks/use-site-metadata';
+import { DOCS_URL } from '../../constants';
+import { assertTrailingSlash } from '../../utils/assert-trailing-slash';
 
 const getBreadcrumbList = (breadcrumb, siteUrl) =>
   breadcrumb.map(({ path, plaintext }, index) => ({
@@ -21,7 +21,7 @@ const BreadcrumbSchema = ({ breadcrumb = [], siteTitle, slug }) => {
       '@type': 'ListItem',
       position: 1,
       name: 'MongoDB Documentation',
-      item: assertTrailingSlash(baseUrl(true)),
+      item: DOCS_URL,
     },
     ...getBreadcrumbList(
       [...(slug !== '/' && project !== 'landing' ? [{ path: '/', plaintext: siteTitle }] : []), ...breadcrumb],
