+++ src/components/Breadcrumbs/index.js
@@ -3,10 +3,9 @@ import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { uiColors } from '@leafygreen-ui/palette';
 import BreadcrumbSchema from './BreadcrumbSchema';
-import { theme } from '../theme/docsTheme';
 import BreadcrumbContainer from './BreadcrumbContainer';
-import { baseUrl } from '../utils/dotcom';
-import { assertTrailingSlash } from '../utils/assert-trailing-slash';
+import { DOCS_URL } from '../../constants';
+import { theme } from '../../theme/docsTheme';
 
 const Wrapper = styled('nav')`
   font-size: ${theme.fontSize.small};
@@ -25,7 +24,7 @@ const Wrapper = styled('nav')`
 const Breadcrumbs = ({ homeUrl = null, pageTitle = null, parentPaths, siteTitle, slug }) => {
   const homeCrumb = {
     title: 'Docs Home',
-    url: homeUrl || assertTrailingSlash(baseUrl(true)),
+    url: homeUrl || DOCS_URL,
   };
   // If a pageTitle prop is passed, use that as the last breadcrumb instead
   const lastCrumb = {
